search-user-list-and-register

+++ src/app/food/food-service.ts
@@ -3,6 +3,7 @@ import {FoodModel} from "./models/food-model";
 import {HttpClient, HttpHeaders} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {environment} from "../../environments/environment";
+import {CommonUtility} from "../utility/common-utility";
 
 @Injectable({
   providedIn:'root'
@@ -15,25 +16,37 @@ export class FoodService{
 
 
   getFoods():Observable<FoodModel.FoodItem[]>{
-    return this.httpClient.get<FoodModel.FoodItem[]>(environment.api + '/food/alllist')
+
+    const jsonString = localStorage.getItem('selectedUser') || '';
+    const userId = CommonUtility.getInstance().getSelectedUserId(jsonString)
+    return this.httpClient.get<FoodModel.FoodItem[]>(environment.api + '/food/alllist?userId=' + userId)
   }
 
   getFoodDetail(id:number):Observable<FoodModel.FoodItem>{
-    return this.httpClient.get<FoodModel.FoodItem>(environment.api + '/food/' + id)
+    const jsonString = localStorage.getItem('selectedUser') || '';
+    const userId = CommonUtility.getInstance().getSelectedUserId(jsonString)
+    return this.httpClient.get<FoodModel.FoodItem>(environment.api + '/food/' + id + '?userId=' + userId)
   }
 
   updateFood(body:any):Observable<FoodModel.FoodItem>{
-    return this.httpClient.post<FoodModel.FoodItem>(environment.api + '/food/createOrUpdate',body)
+    const jsonString = localStorage.getItem('selectedUser') || '';
+    const userId = CommonUtility.getInstance().getSelectedUserId(jsonString)
+    return this.httpClient.post<FoodModel.FoodItem>(environment.api + '/food/createOrUpdate?userId=' + userId,body)
   }
 
   delete(id:number){
     const headers = new HttpHeaders().set('Content-Type', 'text/plain; charset=utf-8');
 
-    return this.httpClient.delete(environment.api + '/food/' + id, { headers, responseType: 'text'});
+    const jsonString = localStorage.getItem('selectedUser') || '';
+    const userId = CommonUtility.getInstance().getSelectedUserId(jsonString)
+
+    return this.httpClient.delete(environment.api + '/food/' + id + '?userId=' + userId, { headers, responseType: 'text'});
   }
 
   search(search:string):Observable<FoodModel.FoodItem[]>{
-    return this.httpClient.get<FoodModel.FoodItem[]>(environment.api + '/food/search/' + search)
+    const jsonString = localStorage.getItem('selectedUser') || '';
+    const userId = CommonUtility.getInstance().getSelectedUserId(jsonString)
+    return this.httpClient.get<FoodModel.FoodItem[]>(environment.api + '/food/search/' + search + '?userId=' + userId)
   }
